consolidate storage directories
